backend/requirements.txt
asgiref==3.10.0
Django==4.2.7
django-cors-headers==4.9.0
djangorestframework==3.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
